StatusEffects (Transferring statusEffectData)

+++ VR Game/Assets/_Sway VR/Health/Scripts/Interface/IDamageable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void DamageHP(float damage);
+    void DamageHP(int damage);
 }
